+++ railties/lib/rails/generators/rails/app/app_generator.rb
@@ -153,7 +153,7 @@ module Rails
     def system_test
       empty_directory_with_keep_file "test/system"
 
-      template "test/system_test_helper.rb"
+      template "test/application_system_test_case.rb"
     end
 
     def tmp
