+++ railties/lib/rails/generators/rails/app/templates/test/application_system_test_case.rb
