+++ railties/lib/rails/generators/rails/plugin/templates/test/application_system_test_case.rb
