+++ railties/lib/rails/generators/test_unit/system/system_generator.rb
@@ -6,8 +6,8 @@ module TestUnit # :nodoc:
       check_class_collision suffix: "Test"
 
       def create_test_files
-        if !File.exist?(File.join("test/system_test_helper.rb"))
-          template "system_test_helper.rb", File.join("test", "system_test_helper.rb")
+        if !File.exist?(File.join("test/application_system_test_case.rb"))
+          template "application_system_test_case.rb", File.join("test", "application_system_test_case.rb")
         end
 
         template "system_test.rb", File.join("test/system", "#{file_name.pluralize}_test.rb")
