+++ railties/lib/rails/generators/test_unit/system/templates/application_system_test_case.rb
