+++ railties/lib/rails/generators/test_unit/system/templates/system_test.rb
@@ -1,4 +1,4 @@
-require "system_test_helper"
+require "application_system_test_case"
 
 class <%= class_name.pluralize %>Test < ApplicationSystemTestCase
   # test "visiting the index" do
